+++ src/Stryker.CLI/Stryker.CLI.UnitTest/ConfigBuilderTests.cs
@@ -2,6 +2,7 @@ using System.IO;
 using McMaster.Extensions.CommandLineUtils;
 using Moq;
 using Shouldly;
+using Stryker.CLI.CommandLineConfig;
 using Stryker.Core.Exceptions;
 using Stryker.Core.Options;
 using Stryker.Core.Options.Inputs;
