+++ src/Stryker.CLI/Stryker.CLI/CommandLineConfig/GroupedHelpTextGenerator.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using System.IO;
 using System.Linq;
+using Stryker.CLI.CommandLineConfig;
 
 namespace Stryker.CLI
 {
